+++ back-end/src/main/java/com/exmaple/flory/util/SecurityUtil.java
@@ -10,13 +10,13 @@ public class SecurityUtil {
     private SecurityUtil() { }
 
     // Request 가 들어올 때 JwtFilter 의 doFilter 에서 저장
-    public static Integer getCurrentMemberId() {
+    public static Long getCurrentMemberId() {
         final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         if (authentication == null || authentication.getName() == null) {
             throw  new RuntimeException("Security Context 에 인증 정보가 없습니다.");
         }
 
-        return Integer.parseInt(authentication.getName());
+        return Long.parseLong(authentication.getName());
     }
 }
